requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
sqlalchemy==2.0.36
pydantic==2.12.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymysql==1.1.1
python-dotenv==1.2.1
